Add token authorization

+++ app/services/project/delete_service.py
@@ -1,11 +1,18 @@
+from boto3.dynamodb.conditions import Key
 from botocore.exceptions import ClientError
-from fastapi.responses import JSONResponse
+from fastapi import HTTPException
 
 from app.core.database import dynamodb
 
 
-def delete_service(project_id: str):
+def delete_service(project_id: str, user):
     table = dynamodb.Table("projects")
+
+    response = table.query(KeyConditionExpression=Key("id").eq(project_id))
+    project_owner = response["Items"][0].get("project_owner")
+
+    if project_owner != user.get("Username"):
+        raise HTTPException(status_code=403, detail="Forbidden")
 
     try:
         response = table.delete_item(
@@ -15,4 +22,4 @@ def delete_service(project_id: str):
         )
         return response
     except ClientError as e:
-        return JSONResponse(content=e.response["Error"], status_code=500)
+        raise HTTPException(detail=e.response["Error"], status_code=500)
